+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+Pkg v1.8 Release Notes
+======================
+
+- New `compat::Bool` kwarg to `Pkg.status` (`--compat` in the REPL mode) to show
+  information about packages not at the latest version.
+
 Pkg v1.7 Release Notes
 ======================